requirements.txt
matplotlib==3.0.0
numpy==1.17.2
PyGithub==1.53
requests==2.23.0
requests-oauthlib==1.3.0
